requirement.txt
alembic==0.9.6
amqp==1.4.9
aniso8601==1.3.0
antiorm==1.2.1
anyjson==0.3.3
asn1crypto==0.23.0
attrs==17.3.0
Automat==0.6.0
beautifulsoup4==4.6.0
billiard==3.3.0.23
blinker==1.4
bs4==0.0.1
celery==3.1.25
celery-with-redis==3.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==2.1.3
cssselect==1.0.1
db==0.1.1
diff-match-patch==20121119
Django==1.11.4
django-celery==3.2.2
django-crispy-forms==1.7.0
django-formtools==2.1
django-import-export==0.6.0
django-reversion==2.0.11
django-tinymce==2.6.0
dominate==2.3.1
et-xmlfile==1.0.1
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Email==1.4.4
Flask-HTTPAuth==3.2.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.5.2
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
future==0.16.0
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
itsdangerous==0.24
jdcal==1.3
Jinja2==2.10
kombu==3.0.37
lxml==4.1.1
Mako==1.0.7
MarkupSafe==1.0
mysqlclient==1.3.12
odfpy==1.3.5
olefile==0.44
openpyxl==2.4.9
parsel==1.2.0
Pillow==4.3.0
pyasn1==0.3.7
pyasn1-modules==0.1.5
pycparser==2.18
PyDispatcher==2.0.5
PyMySQL==0.7.11
pyOpenSSL==17.3.0
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.3
pywin32==221
PyYAML==3.12
queuelib==1.4.2
redis==2.10.6
requests==2.18.4
selenium==3.8.0
service-identity==17.0.0
six==1.11.0
SQLAlchemy==1.1.15
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
utils==0.9.0
vboxapi==1.0
vine==1.1.4
virtualenv==15.1.0
visitor==0.1.3
w3lib==1.18.0
Werkzeug==0.13
WTForms==2.1
xadmin==0.6.1
xlrd==1.1.0
xlwt==1.3.0
zope.interface==4.4.3
